move card container to shared components
